refactor(ui): flatten http model imports
Closes #2482
