+++ apps/ui/src/api/history.ts
@@ -1,6 +1,6 @@
 import { apiJson, apiJsonResponse } from "./http";
 import { cloneTransportValue } from "../transport/clone";
-import type * as Local from "../transport/http_models";
+import type * as Local from "../api/types";
 import type * as Transport from "./types";
 
 export function historyExportUrl(runId: string): string {
