+++ apps/ui/src/app/car_selection_state.ts
@@ -1,4 +1,4 @@
-import type { CarRecord } from "../transport/http_models";
+import type { CarRecord } from "../api/types";
 import type { SettingsState } from "./ui_app_state";
 import { trackAppStateSlice } from "./ui_app_state";
 import { computed, type ReadonlySignal } from "./ui_signals";
