+++ apps/ui/src/app/features/update_feature_workflow.ts
@@ -14,7 +14,7 @@ import type {
   UpdateStartRequestPayload,
   UpdateStatusPayload,
   UsbInternetStatusPayload,
-} from "../../transport/http_models";
+} from "../../api/types";
 import type {
   UpdateFeatureRenderState,
   UpdateFeatureStartIntent,
