+++ apps/ui/src/app/views/settings_speed_source_presenter.ts
@@ -2,7 +2,7 @@ import type {
   ObdDevicePayload,
   ObdStatusPayload,
   SpeedSourceStatusPayload,
-} from "../../transport/http_models";
+} from "../../api/types";
 import {
   deriveDisplayedSpeedSourceMode,
   isManualLikeSpeedSource,
