+++ apps/ui/src/app/views/update_feature_presenter.ts
@@ -3,7 +3,7 @@ import type {
   UpdateStartRequestPayload,
   UpdateStatusPayload,
   UsbInternetStatusPayload,
-} from "../../transport/http_models";
+} from "../../api/types";
 import type {
   InternetPanelRenderModel,
   UpdateTransportChoiceCardRenderModel,
