--- apps/ui/src/transport/http_models.ts
@@ -1,55 +1,0 @@
-export type {
-  AnalysisSettingsPayload,
-  AnalysisSettingsRequest,
-  CarLibraryBrandsPayload,
-  CarLibraryGearbox,
-  CarLibraryModel,
-  CarLibraryModelsPayload,
-  CarLibraryTireOption,
-  CarLibraryTypesPayload,
-  CarLibraryVariant,
-  CarRecord,
-  CarsPayload,
-  CarUpsertRequest,
-  ClientLocationsResponse,
-  DeleteHistoryRunPayload,
-  EspFlashCancelPayload,
-  EspFlashHistoryAttemptPayload,
-  EspFlashHistoryPayload,
-  EspFlashLogsPayload,
-  EspFlashPortsPayload,
-  EspFlashStartPayload,
-  EspFlashStatusPayload,
-  EspSerialPortPayload,
-  FindingPayload,
-  HealthStatusPayload,
-  HistoryEntry,
-  HistoryInsightWarningPayload,
-  HistoryInsightsAnalyzingPayload,
-  HistoryInsightsPayload,
-  HistoryInsightsResult,
-  HistoryListPayload,
-  HistoryRunPayload,
-  HistoryRunAnalysisPayload,
-  HistoryRunAnalysisFindingPayload,
-  HistoryRunPeakTableRowPayload,
-  HistoryRunPlotPayload,
-  LanguagePayload,
-  LocationOption,
-  LoggingStatusPayload,
-  ObdDevicePayload,
-  ObdPairPayload,
-  ObdScanPayload,
-  ObdStatusPayload,
-  SpeedSourceKind,
-  SpeedSourcePayload,
-  SpeedSourceRequest,
-  SpeedSourceStatusPayload,
-  SpeedUnitPayload,
-  UpdateCancelPayload,
-  UpdateIssue,
-  UpdateStartPayload,
-  UpdateStartRequestPayload,
-  UpdateStatusPayload,
-  UsbInternetStatusPayload,
-} from "../api/types";
