+++ apps/ui/tests/settings_speed_source_presenter.spec.ts
@@ -4,7 +4,7 @@ import type {
   ObdDevicePayload,
   ObdStatusPayload,
   SpeedSourceStatusPayload,
-} from "../src/transport/http_models";
+} from "../src/api/types";
 import {
   buildSettingsSpeedSourcePanelModel,
   buildSpeedSourceDiagnosticsRenderModel,
